+++ docs/docs/CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Static asset builds are now done on demand instead of the results being committed to source control
 - The Dockerfile has been removed as it is no longer in use
 - Developer documentation has been added to the docs site
+- Added a header 'X-Target-Backend' to redirect the request to a different backend/target than the TARGET (default is TARGET)
 
 ## v1.15.0